Implement on_remove_user_third_party_identifier
This method will watch for third-party identifiers being
removed from a user's account on Synapse, and subsequently
unbind them from the configured Sydent instance.

We no longer need to inform the homeserver of the binding, as the module
will be able to perform the unbinding now.

Having the homeserver perform the unbind was flawed, as homeservers will
only attempt to talk to identity servers over HTTPS,
whereas this module supports Sydent instances available only over HTTP as well.

The only drawback from this change is that the bind will no longer be removed
if this module is uninstalled. But that doesn't seem particularly unexpected.